Fix map URL position error caused by legend removal
`map.invalidateSize()` moves the map to the default map position, ignoring the leaflet-hash URL (which does not even receive the position from the hash initially).

To work around this, we set the default-position to the leaflet-hash URL position if present.

Other fixes:
- switching the map via the primary navigation does not reload anymore but preserves the map position. It also hides/shows the legend as needed.

(Switching the layer with the layer control still shows the legend even if empty, which is good in the sense that the layer control does not change position this way…)

- the legend is only visible on the parkingmap but shown every time the parking map is selected